+++ R/deriv.R
@@ -1,3 +1,59 @@
+#' Calculate time since melt from transitions data frame #'
+#' @param date Date: one or more dates of interest 
+#' @param trx : the transitions, in the form of a data.frame, string (giving the file name of the database file), or a connection to the database file #'
+#' @return A raster stack, with one layer per date 
+#' @noRd
+#' @keywords internal
+calc_time_since_melt <- function(date, trx) {
+  psproj <- "+proj=stere +lat_0=-90 +lat_ts=-70 +lon_0=0 +x_0=0 +y_0=0 +a=6378273 +b=6356889.449 +units=m +no_defs"
+  ice_template <- raster::raster(nrows = 332, ncols = 316, crs = psproj)
+  raster::extent(ice_template) <- c(-3950000, 3950000, -3950000, 4350000)
+  
+  date <- as.Date(date)
+  if (is.character(trx)) {
+    dbh <- DBI::dbConnect(duckdb::duckdb(dbdir = trx, read_only = TRUE))
+    trx <- dplyr::collect(dplyr::tbl(dbh, "transitions")) ## get the whole table
+    DBI::dbDisconnect(dbh, shutdown = TRUE) ## close after use if we were given a filename, but not if we were given the db handle directly
+  } else if (inherits(trx, "DBIConnection")) {
+    trx <- dplyr::collect(dplyr::tbl(trx, "transitions"))
+  }
+  ## pre-sort it by idx and date (ascending)
+  trx <- trx[order(trx$idx, trx$date, decreasing = FALSE), ]
+  
+  ## reconstruct land mask
+  land_idx <- setdiff(seq_len(332 * 316), unique(trx$idx[!is.na(trx$state)]))
+  
+  out <- lapply(date, function(d) {
+    ## find most recent state of each cell (prior to or at d)
+    this <- trx[trx$date <= d, ]
+    this <- this[c(which(diff(this$idx) > 0), nrow(this)), ]
+    ## that's the last entry prior to or at d
+    ## but for cells that are still frozen at date d, when did they next melt? for this, want the first entry after d
+    this2 <- trx[trx$date > d, ]
+    this2 <- this2[c(1L, which(diff(this2$idx) > 0) + 1L), ]
+    time_since_melt <- ice_template
+    tv <- rep(NA_integer_, prod(dim(time_since_melt)))
+    melted_idx <- which(this$state == -1)
+    tv[this$idx[melted_idx]] <- d - this$date[melted_idx]
+    ## open water, never frozen
+    ow_idx <- this$idx[which(this$state == -2)]
+    tv[ow_idx] <- 2^15 - 3L ## 32765 = open ocean zone
+    melt_later_idx <- which(this2$state == -1 & !this2$idx %in% this$idx[melted_idx]) ## i.e. melted in future, not already melted
+    tv[this2$idx[melt_later_idx]] <- d - this2$date[melt_later_idx]
+    pf_idx <- this$idx[which(this$state >= 1 & !this$idx %in% this2$idx[melt_later_idx])] ## permanently frozen over the full data record
+    tv[pf_idx] <- -2^15 ## -32768 = never-melted ice
+    ##tv[this$idx[is.na(this$state)]] <- 199 ## no transitions observed
+    tv[intersect(land_idx, which(is.na(tv)))] <- 2^15 - 2L ## 32766 = land
+    ## previously we used 32767 = missing data, NA, when exporting to netcdf
+    raster::values(time_since_melt) <- tv
+    time_since_melt
+  })
+  raster::stack(out)
+}
+
+
+
+
 
 ##' Read data from derived sea ice data products.
 ##'
@@ -60,41 +116,12 @@ readderivice <- function(date,
   if (missing(date)) {
     if (latest) date <- max(files$date)  else date <- min(files$date)
   }
+
   date <- timedateFrom(date)
   files <- .processFiles(date, files, time.resolution)
   
-  ## check that files are available
-  
-  
-  ## NSIDC projection and grid size for the Southern Hemisphere
-  prj  <- "+proj=stere +lat_0=-90 +lat_ts=-70 +lon_0=0 +k=1 +x_0=0 +y_0=0 +a=6378273 +b=6356889.449 +units=m +no_defs " 
-  cropit <- FALSE
-  if (!is.null(xylim)) {
-    cropit <- TRUE
-    cropext <- extent(xylim)
-  }
-  
-  nfiles <- nrow(files)
-  ## progress
-  if (nrow(files) > 1L) {
-  pb <- progress::progress_bar$new(
-    format = "  extracting [:bar] :percent in :elapsed",
-    total = nfiles, clear = FALSE, width= 60)
-  pb$tick(0)
-  }
-  r <- vector("list", nfiles)
-  ## loop over file indices
-  for (ifile in seq_len(nfiles)) {
-    r0 <- raster(files$fullname[ifile])
-    if (cropit) r0 <- crop(r0, cropext)
-    r[[ifile]] <- r0
-    if (nrow(files) > 1L) { pb$tick()}
-  }
-  if (nfiles > 1) r <- brick(stack(r), ...) else r <- r[[1L]]
-  
-  projection(r) <- prj
-  names(r) <- basename(files$file)
-  setZ(r, files$date)
+  out <- raster::stack(lapply(files$date, calc_time_since_melt, trx = cfiles0))
+  setZ(out, files$date)
 }
 
 
@@ -110,30 +137,17 @@ readderivice <- function(date,
 ##' @return data.frame of \code{file} and \code{date}
 derivicefiles <- function(product = "time_since_melt", ...) {
   
-  datadir <- getOption("default.datadir")
-  product <- match.arg(product)
-  ftx <- .allfilelist()
+ product <- match.arg(product)
+ ftx <- raadfiles::get_raad_filenames(all = TRUE)
+idx0 <- grep("smmr_ssmi_nasateam/time_since_melt/time_since_melt.duckdb", ftx$file)
+ cfiles0 <- file.path(ftx$root[idx0], ftx$file[idx0])
  
- cfiles0 <- grep("webdav.data.aad.gov.au", ftx, value = TRUE)
-  cfiles <- grep("time_since_melt", cfiles0, value = TRUE)
- 
-  if (length(cfiles) < 1) stop("no files found")
-  
-  doffs <- 0
-  sep <- "_"
-  datepart <- sapply(strsplit(basename(cfiles), sep), function(x) x[length(x) - doffs])
-  
-  datepattern <-  "%Y%m%d"
-  dates <- timedateFrom(as.POSIXct(strptime(datepart, datepattern, tz = "GMT")))
-  
-  nas <- is.na(dates)
-  dates <- dates[!nas]
-  cfiles <- cfiles[!nas]
-  
-  cfs <- data.frame(file = gsub(paste(datadir, "/", sep = ""), "", cfiles), date = dates,
-                    fullname = cfiles, stringsAsFactors = FALSE)[order(dates), ]
-  
-  cfs <- cfs[!duplicated(cfs$date), ]
-  
-  cfs
+if (length(cfiles0) < 1 || !file.exists(cfiles0)) stop("no time_since_melt resource found")
+
+ dbh <- DBI::dbConnect(duckdb::duckdb(dbdir = cfiles0, read_only = TRUE))
+ trx <- dplyr::collect(dplyr::arrange(dplyr::distinct(dplyr::tbl(dbh, "transitions"), date), date))
+ trx$date <- as.POSIXct(trx$date, tz = "UTC")
+ DBI::dbDisconnect(dbh, shutdown = TRUE) ## close after use if we were given a filename, but not if we were given the db handle directly
+ trx$fullname <- cfiles0
+ trx
 }