+++ NEWS.md
@@ -1,5 +1,7 @@
 # raadtools dev
 
+* Replace old form of time_since_melt with fancy new duckdb process. 
+
 * Fixed matching files bug for 8 day intervals. 
 
 * Fixed readCHL_month() to work with extract.